feat: IsoError.trace() prints module name.
Move all code into a single file to reduce overheads.

This is done so because the library is small.
In normal case they of course should be in different files.
