--- src/ModuleError.spec.ts
@@ -1,8 +1,0 @@
-import a from 'assertron';
-import { ModuleError } from '.';
-
-test('with module', () => {
-  const e = new ModuleError('iso-error', 'mod')
-
-  a.satisfies(e, { module: 'iso-error' })
-})
